Move functionality to create file tree into its own file
